langgraph_deep_research/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
parallel-web>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
